Create Record Death from HCN Alert.vbs
The 'Record Death from HCN Alert' script allows the user to select a Date of Death Alert email from the Health and Care Index, press the key that the script is mapped to on the CRT terminal emulator, and the script will get the H+C number from the email. Using this H+C number it will look up the patient on CLINiCOM/PAS, compare the PMI (Patient Master Index) record side by side with CLINiCOM's HCN interface, and from there it will retrieve the date of death held by the HCN index. The script will then open the Record Patient Death function, look up the patient, insert the date of death and populate the informed-by field and then await the user to press Y in the "Enter?" field.