Added description comment and renamed
Added a comment describing the function of the script as it is stored in a shared folder for other users to access and the only description was the file name. Now anyone who opens the file in a text editor will be able to understand the function. Also renamed the file from "Trim Phone Numbers.vbs" to "Validate Phone Numbers". This name describes the script a bit better.