Added desc and ability to dispose of email
Added a description and instructions at the start of the script so that those who open the file in a text editor can use it properly. The script will now also mark the email as read and move it to a separate folder once it detects that the <kbd>Y</kbd> key is pressed.