Added predicted next start time when adding
When adding a new timeslot the script will autofill the next timeslot's Start Time with the previous slot's End Time. This will be accurate when timeslots are back to back (85% of the time) however sometimes there are breaks between timeslots, where the user just needs to backspace over the predicted time.