Change "Extract" to "Summary" in name for accuracy
I changed the file name from `Masterfile Data Extract` to `Master File Data Summary`. This is required to make room for a new script which will assume the previous name and to allow both scripts to be named accurately. At the end of the day this script only provides a record's code and description, rather than all record details.