added Function navigation methods and GetSubtitle
Added GoToField(index) to try to jump straight to the prompt/field at that index.
Added ResetForm() which attempts to go back to the very first field in the function and clears it, effectively resetting all the prompts/fields.
Added GetSubtitle() in order to help identify which subfunction, if any, is currently in use. This is useful for understanding what part of a function is currently open and allows scripts to be orientated.