removed GoToFieldOccurrence check for prompt
Removed the check for the Occurrence No. prompt as it does not appear in the same place every time. The Go To Field "Window" location depends on where the terminal cursor is. It tests OK at the minute without the check, so it's not quite worth adding unless we run into issues.